Adds #unshift to hash
Hash#shift removes the first key-value pair, returning them as an array

ex.
```
{item: "Hot Dog", price: 3.5, denomination: "USD"}.shift
[:item, "Hot Dog"]
```

Ruby includes no #unshift for hashes (although it does for Arrays) which is
unfair.

This PR adds Hash#unshift which takes a key-value pair as an array and adds it
to the hash in the "beginning" position:

ex.
```
{price: 3.5, denomination: "USD"}.unshift [:item, "Hot Dog"]
{item: "Hot Dog", price: 3.5, denomination: "USD"}
```